transform/news/news_issue_attention/requirements.txt
pandas
python-dotenv
openai
psycopg2-binary
sqlalchemy